requirements.txt
vosk
ffmpeg-python
wxpython